[#119] rename: 알림 서버 통신 관련 패키지명 수정
