+++ main/src/main/java/org/sopt/makers/crew/main/common/config/FeignClientConfig.java
@@ -3,7 +3,7 @@ package org.sopt.makers.crew.main.common.config;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Configuration;
 
-@EnableFeignClients(basePackages = "org.sopt.makers.crew.main.external.notification")
+@EnableFeignClients(basePackages = "org.sopt.makers.crew.main.internal.notification")
 @Configuration
 public class FeignClientConfig {
